pyproject.toml
[project]
name = "embedding-from-scratch"
version = "0.1.0"
description = "Entraîner un modèle d'embedding — code source du livre"
readme = "README.md"
authors = [
    { name = "antoine126" }
]
requires-python = ">=3.12"
dependencies = [
    "torch>=2.5.1",
    "transformers>=4.47.0",
    "tokenizers>=0.21.0",
    "datasets>=3.2.0",
    "sentence-transformers>=3.3.0",
    "faiss-cpu>=1.9.0",
    "numpy>=2.1.0",
    "scipy>=1.14.0",
    "scikit-learn>=1.6.0",
    "wandb>=0.19.0",
    "loguru>=0.7.2",
    "rich>=13.9.0",
    "fastapi>=0.115.0",
    "uvicorn>=0.32.0",
    "psycopg2-binary>=2.9.10",
    "rank-bm25>=0.2.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-cov>=6.0.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
    "ipykernel>=6.29.0",
    "jupyter>=1.1.0",
]
eval = [
    "mteb>=1.31.0",
    "beir>=2.0.0",
]

[tool.uv.sources]
torch = [
  { index = "pytorch-cu128"},
]
torchvision = [
  { index = "pytorch-cu128"},
]

[[tool.uv.index]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true

[[tool.uv.index]]
name = "pytorch-cu128"
url = "https://download.pytorch.org/whl/cu128"
explicit = true

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/embedding"]

[tool.ruff]
line-length = 88
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP"]

[tool.mypy]
python_version = "3.12"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--cov=src/embedding --cov-report=term-missing"
